+++ client/src/components/list/ListElement.tsx
@@ -1,12 +1,14 @@
 import React from 'react';
 import {StyleSheet, Text, TouchableOpacity, View} from 'react-native';
 import {IconButton} from 'react-native-paper';
+import Swipeable from 'react-native-swipeable-row';
 
 interface Props {
     _id,
     title,
     inProgress,
     onEdit,
+    onDelete,
     changeProgress,
 }
 
@@ -16,23 +18,51 @@ const ListElement = (props: Props) => {
     title,
     inProgress,
     onEdit,
+    onDelete,
     changeProgress,
   } = props;
 
-  return(
-  <View style={styles.rowContainer}>
-    <IconButton
-        icon={inProgress ? "circle-outline" : "check"}
-        size={30}
-        onPress={() => {changeProgress(_id, inProgress)}}
-    />
+    const leftContent = [
+        <View style={styles.leftSwipeContener}>
+            <Text style={styles.leftSwipeItemText}>Delete</Text>
+        </View>
+    ];
 
-    <TouchableOpacity style={styles.content}
-                      onPress={() => {onEdit(_id)}}
-    >
-        <Text style={[!inProgress ? styles.crossOver : styles.text]}>{title}</Text>
-    </TouchableOpacity>
-  </View>
+  return(
+      inProgress ? (
+          <View style={styles.rowContainer}>
+              <IconButton
+                  icon={inProgress ? "circle-outline" : "check"}
+                  size={30}
+                  onPress={() => {changeProgress(_id, inProgress)}}
+              />
+
+              <TouchableOpacity style={styles.content}
+                                onPress={() => {onEdit(_id)}}
+              >
+                  <Text style={[!inProgress ? styles.crossOver : styles.text]}>{title}</Text>
+              </TouchableOpacity>
+          </View>
+      ) : (
+          <Swipeable leftContent={leftContent}
+                     leftActionActivationDistance={200}
+                     onLeftActionComplete={() => onDelete(_id)}
+          >
+              <View style={styles.rowContainer}>
+                  <IconButton
+                      icon={inProgress ? "circle-outline" : "check"}
+                      size={30}
+                      onPress={() => {changeProgress(_id, inProgress)}}
+                  />
+
+                  <TouchableOpacity style={styles.content}
+                                    onPress={() => {onEdit(_id)}}
+                  >
+                      <Text style={[!inProgress ? styles.crossOver : styles.text]}>{title}</Text>
+                  </TouchableOpacity>
+              </View>
+          </Swipeable>
+      )
   );
 };
 
@@ -53,7 +83,16 @@ const styles = StyleSheet.create({
     },
     text:{
 
-    }
+    },
+    leftSwipeContener: {
+        height: 50,
+        alignItems: 'flex-end',
+        justifyContent: 'center',
+        backgroundColor: '#ff3232',
+    },
+    leftSwipeItemText:{
+        paddingRight: 20,
+    },
 });
 
 export default ListElement;
